move EventLog to types and fix not deal with clientRegisterEventLog response
