--- core/types/event_log_params.go
@@ -1,10 +1,0 @@
-package types
-
-type EventLogParams struct {
-	FromBlock string   `json:"fromBlock"`
-	ToBlock   string   `json:"toBlock"`
-	Addresses []string `json:"addresses"`
-	Topics    []string `json:"topics"`
-	GroupID   string   `json:"groupID"`
-	FilterID  string   `json:"filterID"`
-}
